Rearrange files + other minor changes
