+++ _templates/route/new/index.ejs.t
@@ -1,11 +1,10 @@
 ---
-to: <%- filepath %>/index.tsx
-sh: prettier --write <%- filepath %>/index.tsx
+to: <%- filepath %>.tsx
+sh: prettier --write <%- filepath %>.tsx
 ---
 
 import { h } from 'preact'
 import Page from '@/components/page'
-import style from './style'
 
 interface Props {
 <% props.forEach(prop => { _%>
@@ -14,7 +13,7 @@ interface Props {
 }
 
 const <%- name %> = ({<%- props.join(', ') %>}: Props) => (
-  <Page name="<%- name %>">
+  <Page name="<%- name %>" back={false}>
     <h1><%- name %> - {JSON.stringify({<%- props.join(', ') %>})}</h1>
   </Page>
 )
