+++ _templates/route/new/router.ejs.t
@@ -1,10 +1,11 @@
 ---
-to: src/index.tsx
+to: src/app.tsx
 inject: true
-before: </Router>
+before: ]}
 sh: prettier --write src/index.tsx
 ---
-    <Route<{ <%- props.map(p => p + ': string ').join(';') %> }>
-      path="<%- url %>"
-      component={asyncRoute(() => import('./<%- filepath.replace(/^src\//, '') %>'))}
-    />
+
+          {
+            path: '<%- url %>',
+            component: () => import('./<%- filepath.replace(/^src\//, '),
+          },
