--- _templates/route/new/style.ejs.t
@@ -1,3 +1,0 @@
----
-to: <%- filepath %>/style.css
----
