+++ package.json
@@ -39,7 +39,7 @@
   "dependencies": {
     "matchit": "^1.0.8",
     "polished": "^3.3.0",
-    "preact": "^10.0.0-beta.1",
+    "preact": "^10.0.0-beta.2",
     "systemjs": "3.1.6"
   },
   "jest": {
@@ -112,7 +112,9 @@
     "prettier": "1.17.1",
     "prompts": "1.2.1",
     "rollup": "1.13.1",
+    "rollup-plugin-alias": "^1.5.2",
     "rollup-plugin-babel": "4.3.2",
+    "rollup-plugin-commonjs": "^10.0.0",
     "rollup-plugin-node-resolve": "5.0.1",
     "rollup-plugin-postcss": "2.0.3",
     "rollup-plugin-terser": "5.0.0",
