+++ src/index.tsx
@@ -1,53 +1,6 @@
-import './style.css'
-import { useRouter } from './router'
-import { h, render, Fragment } from 'preact'
-import 'preact/debug'
+import App from './app'
 import { cleanupTokens } from './jwt'
-import { DialogDisplayer } from './components/dialog'
-
-const App = () => (
-  <Fragment>
-    {useRouter([
-      {
-        path: '/',
-        component: () => import('./routes/home'),
-      },
-      {
-        path: '/events/:eventKey',
-        component: () => import('./routes/event'),
-      },
-      {
-        path: '/events/:eventKey/matches/:matchKey',
-        component: () => import('./routes/event-match'),
-      },
-      {
-        path: '/events/:eventKey/matches/:matchKey/scout',
-        component: () => import('./routes/scout'),
-      },
-      {
-        path: '/events/:eventKey/teams/:teamNum',
-        component: () => import('./routes/event-team'),
-      },
-      {
-        path: '/users',
-        component: () => import('./routes/users'),
-      },
-      {
-        path: '/register',
-        component: () => import('./routes/register'),
-      },
-      {
-        path: '/login',
-        component: () => import('./routes/login'),
-      },
-      {
-        path: '/leaderboard',
-        component: () => import('./routes/leaderboard'),
-      },
-    ])}
-    <DialogDisplayer />
-  </Fragment>
-)
+import { render, h } from 'preact'
 
 if (process.env.NODE_ENV === 'development') {
   if (module.hot) {
